[#12] Feat: 소비기록 연관관계 매핑 설정
