+++ src/main/java/com/server/money_touch/domain/notification/entity/NotificationType.java
@@ -1,0 +1,23 @@
+package com.server.money_touch.domain.notification.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@EntityListeners(AuditingEntityListener.class)
+public class NotificationType {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, length = 255)
+    private String imgUrl;
+
+    @Column(nullable = false, length = 10)
+    private String notificationTypeName;
+}
