+++ src/main/java/com/server/money_touch/domain/user/entity/AuthType.java
@@ -1,0 +1,5 @@
+package com.server.money_touch.domain.user.entity;
+
+public enum AuthType {
+    LOCAL, SOCIAL
+}
