--- src/main/java/com/server/money_touch/user/entity/Gender.java
@@ -1,5 +1,0 @@
-package com.server.money_touch.user.entity;
-
-public enum Gender {
-    MALE, FEMALE
-}
