--- src/main/java/com/server/money_touch/user/entity/Role.java
@@ -1,5 +1,0 @@
-package com.server.money_touch.user.entity;
-
-public enum Role {
-    ADMIN, USER;
-}
