--- src/main/java/com/server/money_touch/user/entity/SocialLogin.java
@@ -1,32 +1,0 @@
-package com.server.money_touch.user.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.springframework.data.annotation.CreatedDate;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-public class SocialLogin {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String KakaoKey;
-
-    @CreatedDate
-    private LocalDateTime connectedAt;
-
-    // 회원-소셜로그인 일대일
-    @OneToOne
-    @JoinColumn(name = "user_id")
-    private User user;
-}
