--- src/main/java/com/server/money_touch/user/entity/UserDetail.java
@@ -1,38 +1,0 @@
-package com.server.money_touch.user.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-public class UserDetail {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String age;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Gender gender;
-
-    @Column(nullable = false)
-    private String job;
-
-    @Column(nullable = false)
-    private Boolean isIncome;
-
-    // 회원-회원상세 일대일
-    @OneToOne
-    @JoinColumn(name = "user_id")
-    private User user;
-
-}
